requirements.txt
chromadb
sentence-transformers
numpy
pandas
python-dotenv
openai  # Compatible with DeepSeek APIs
